control-plane/pyproject.toml
[project]
name = "alphaconsole-control-plane"
version = "0.1.0"
description = "The permanent orchestration brain for the AlphaConsole platform."
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "psycopg2-binary",
    "nats-py",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "httpx",
    "websockets"
]

[tool.poetry]
name = "alphaconsole-control-plane"
version = "0.1.0"
description = ""
authors = ["AlphaConsole Team"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
